+++ modules/sde-core/src/main/java/org/eclipse/tractusx/sde/core/processreport/ProcessReportUseCase.java
@@ -24,13 +24,10 @@ package org.eclipse.tractusx.sde.core.processreport;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 import org.eclipse.tractusx.sde.common.entities.Policies;
-import org.eclipse.tractusx.sde.common.entities.UsagePolicies;
 import org.eclipse.tractusx.sde.common.enums.ProgressStatusEnum;
-import org.eclipse.tractusx.sde.common.enums.UsagePolicyEnum;
 import org.eclipse.tractusx.sde.core.failurelog.mapper.FailureLogMapper;
 import org.eclipse.tractusx.sde.core.failurelog.repository.FailureLogRepository;
 import org.eclipse.tractusx.sde.core.processreport.entity.ProcessReportEntity;
