+++ modules/sde-core/src/main/java/org/eclipse/tractusx/sde/core/service/SubmodelOrchestartorService.java
@@ -20,7 +20,6 @@
 
 package org.eclipse.tractusx.sde.core.service;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
