+++ modules/sde-core/src/test/java/org/eclipse/tractusx/sde/service/ConsumerControlPanelServiceTest.java
@@ -92,7 +92,7 @@ class ConsumerControlPanelServiceTest {
 
 	@MockBean
 	private ContractOfferRequestFactory contractOfferRequestFactory;
-	
+
 	@MockBean
 	private EDRRequestHelper eDRRequestHelper;
 
@@ -103,8 +103,8 @@ class ConsumerControlPanelServiceTest {
 
 		when(contractOfferCatalogApi.getContractOffersCatalog((JsonNode) any())).thenReturn(json);
 
-		List<QueryDataOfferModel> queryOnDataOffers = consumerControlPanelService
-				.queryOnDataOffers("https://example.org/example", 0, 0, null);
+		List<QueryDataOfferModel> queryOnDataOffers = consumerControlPanelService.queryOnDataOffers("example", "", "",
+				0, 0);
 		assertTrue(queryOnDataOffers.isEmpty());
 		verify(contractOfferCatalogApi).getContractOffersCatalog((JsonNode) any());
 	}
@@ -112,18 +112,10 @@ class ConsumerControlPanelServiceTest {
 	@Test
 	void testQueryOnDataOffersWithUsagePolicies() throws Exception {
 
-		String filterExpression = String.format("""
-				 "filterExpression": [{
-				    "operandLeft": "https://w3id.org/edc/v0.0.1/ns/type",
-				    "operator": "=",
-				    "operandRight": "data.core.digitalTwinRegistry"
-				}]""");
-
 		JsonNode contractOffersCatalogResponse = getCatalogResponse();
 		when(contractOfferCatalogApi.getContractOffersCatalog((JsonNode) any()))
 				.thenReturn(contractOffersCatalogResponse);
-		assertEquals(1, consumerControlPanelService
-				.queryOnDataOffers("https://example.org/example", 0, 1, filterExpression).size());
+		assertEquals(1, consumerControlPanelService.queryOnDataOffers("example", "", "", 0, 1).size());
 		verify(contractOfferCatalogApi).getContractOffersCatalog((JsonNode) any());
 	}
 
@@ -131,9 +123,7 @@ class ConsumerControlPanelServiceTest {
 	void testSubscribeDataOffers1() {
 		ArrayList<Offer> offerRequestList = new ArrayList<>();
 		List<Policies> usagePolicies = new ArrayList<>();
-		Policies usagePolicy = Policies.builder()
-				.technicalKey("BusinessPartnerNumber")
-				.value(List.of("BPN123456789"))
+		Policies usagePolicy = Policies.builder().technicalKey("BusinessPartnerNumber").value(List.of("BPN123456789"))
 				.build();
 		usagePolicies.add(usagePolicy);
 		ConsumerRequest consumerRequest = new ConsumerRequest("42", "https://example.org/example", offerRequestList,
@@ -149,7 +139,6 @@ class ConsumerControlPanelServiceTest {
 		list.addProperty("odrl:and", constraintRequest);
 		when(policyConstraintBuilderService.getUsagePoliciesConstraints(List.of())).thenReturn(list);
 		assertEquals(usagePolicies, policies);
-		assertEquals(1, consumerControlPanelService.getAuthHeader().size());
 	}
 
 	private JsonNode getCatalogResponse() throws JsonProcessingException, JsonMappingException {
